rename app.py to avoid confusion
